requirements.txt
pykwalify
PyYAML